+++ test/OpenTelemetry.Instrumentation.AspNetCore.Tests/BasicTests.cs
@@ -629,18 +629,18 @@ public sealed class BasicTests
     }
 
     [Theory]
-    [InlineData("CONNECT", "CONNECT")]
-    [InlineData("DELETE", "DELETE")]
-    [InlineData("GET", "GET")]
-    [InlineData("PUT", "PUT")]
-    [InlineData("HEAD", "HEAD")]
-    [InlineData("OPTIONS", "OPTIONS")]
-    [InlineData("PATCH", "PATCH")]
-    [InlineData("Get", "GET")]
-    [InlineData("POST", "POST")]
-    [InlineData("TRACE", "TRACE")]
-    [InlineData("CUSTOM", "_OTHER")]
-    public async Task HttpRequestMethodIsSetAsPerSpec(string originalMethod, string expectedMethod)
+    [InlineData("CONNECT", "CONNECT", null)]
+    [InlineData("DELETE", "DELETE", null)]
+    [InlineData("GET", "GET", null)]
+    [InlineData("PUT", "PUT", null)]
+    [InlineData("HEAD", "HEAD", null)]
+    [InlineData("OPTIONS", "OPTIONS", null)]
+    [InlineData("PATCH", "PATCH", null)]
+    [InlineData("Get", "GET", "Get")]
+    [InlineData("POST", "POST", null)]
+    [InlineData("TRACE", "TRACE", null)]
+    [InlineData("CUSTOM", "_OTHER", "CUSTOM")]
+    public async Task HttpRequestMethodIsSetAsPerSpec(string originalMethod, string expectedMethod, string expectedOriginalMethod)
     {
         var exportedItems = new List<Activity>();
 
@@ -681,18 +681,8 @@ public sealed class BasicTests
 
         var activity = exportedItems[0];
 
-        Assert.Contains(activity.TagObjects, t => t.Key == SemanticConventions.AttributeHttpRequestMethod);
-
-        if (originalMethod.Equals(expectedMethod, StringComparison.OrdinalIgnoreCase))
-        {
-            Assert.DoesNotContain(activity.TagObjects, t => t.Key == SemanticConventions.AttributeHttpRequestMethodOriginal);
-        }
-        else
-        {
-            Assert.Equal(originalMethod, activity.GetTagValue(SemanticConventions.AttributeHttpRequestMethodOriginal) as string);
-        }
-
-        Assert.Equal(expectedMethod, activity.GetTagValue(SemanticConventions.AttributeHttpRequestMethod) as string);
+        Assert.Equal(expectedMethod, activity.GetTagValue(SemanticConventions.AttributeHttpRequestMethod));
+        Assert.Equal(expectedOriginalMethod, activity.GetTagValue(SemanticConventions.AttributeHttpRequestMethodOriginal));
     }
 
     [Fact]